+++ components/Toolbar.vue
@@ -1,32 +1,31 @@
 <template>
-  <!-- <div class="nav">  -->
-      <b-navbar toggleable="md" type="dark" variant="info" class='navbar1'>
-        <b-navbar-toggle target="nav_collapse" toggleable></b-navbar-toggle>
-        <b-navbar-brand href="/">LeaveDay</b-navbar-brand>
-        <b-collapse is-nav id="nav_collapse">
-          <b-navbar-nav>
-            <b-nav-item href="/todoList">TodoList</b-nav-item>
-            <b-nav-item href="/jobDiary">JobDiary</b-nav-item>
-          </b-navbar-nav>
-          <!-- 로그인 -->
-          <b-navbar-nav v-if='isAuthenticated' class="ml-auto">
-            <b-nav-item-dropdown right>
-              <template slot="button-content">
-                <em>User</em>
-              </template>
-              <b-dropdown-item href="#">프로필</b-dropdown-item>
-              <b-dropdown-item v-b-modal.modal1 >리스트</b-dropdown-item>
-              <b-dropdown-item href="#" @click="onLogout" >로그아웃</b-dropdown-item>
-            </b-nav-item-dropdown>
-          </b-navbar-nav>
-          <!-- 비로그인 -->
-          <b-navbar-nav v-else class="ml-auto">
-            <b-nav-item href="login">login</b-nav-item>
-          </b-navbar-nav>
-        </b-collapse>
-      </b-navbar>
-      <!-- <leavecreate :userInfo="userInfo"/> -->
-  <!-- </div> -->
+  <b-navbar toggleable="md" type="dark" variant="info" class='navbar1'>
+    <b-navbar-toggle target="nav_collapse"></b-navbar-toggle>
+    <!-- <b-navbar-toggle target="nav_collapse" toggleable></b-navbar-toggle> -->
+    <b-navbar-brand href="/">LeaveDay</b-navbar-brand>
+    <b-collapse is-nav id="nav_collapse">
+      <b-navbar-nav>
+        <b-nav-item href="/todoList">TodoList</b-nav-item>
+        <b-nav-item href="/jobDiary">JobDiary</b-nav-item>
+      </b-navbar-nav>
+      <!-- 로그인 -->
+      <b-navbar-nav v-if='isAuthenticated' class="ml-auto">
+        <b-nav-item-dropdown right>
+          <template slot="button-content">
+            <em>User</em>
+          </template>
+          <b-dropdown-item href="#">프로필</b-dropdown-item>
+          <b-dropdown-item v-b-modal.modal1 >리스트</b-dropdown-item>
+          <b-dropdown-item href="#" @click="onLogout" >로그아웃</b-dropdown-item>
+        </b-nav-item-dropdown>
+      </b-navbar-nav>
+      <!-- 비로그인 -->
+      <b-navbar-nav v-else class="ml-auto">
+        <b-nav-item href="login">login</b-nav-item>
+      </b-navbar-nav>
+    </b-collapse>
+  </b-navbar>
+  <!-- <leavecreate :userInfo="userInfo"/> -->
 </template>
 
 <script>
@@ -66,6 +65,8 @@
     .navbar1 {
       position: fixed;
       top: 0;
-      width: 100%
+      width: 100%;
+      z-index: 1;
     }
+    
 </style>
