+++ package.json
@@ -29,7 +29,7 @@
     "jwt-decode": "^2.2.0",
     "nuxt": "^1.0.0",
     "vue-fullcalendar": "^1.0.9",
-    "vue-material": "^1.0.0-beta-8",
+    "vue-material": "^1.0.0-beta-10.2",
     "vuejs-datepicker": "^1.3.0"
   },
   "devDependencies": {
