+++ plugins/vue-material.js
@@ -1,4 +1,5 @@
 import Vue from 'vue'
 import VueMaterial from 'vue-material'
-import 'vue-material/dist/vue-material.css'
+import 'vue-material/dist/vue-material.min.css'
+
 Vue.use(VueMaterial)
